first refactor of viur scriptor workflow
