+++ setup.cfg
@@ -25,6 +25,8 @@ install_requires =
     pipfile-requirements==0.3.0
     requests==2.31.0
     semver==3.0.2
+    viur-scriptor-api==0.0.3
+
