+++ src/viur_cli/__init__.py
@@ -6,6 +6,6 @@ from .local import *
 from .build import *
 from .setup import *
 from .version import *
-from .scriptor import script
 from .tool import *
 from .update import *
+from .script import script
