--- src/viur_cli/scriptor/__init__.py
@@ -1,1 +1,0 @@
-from .cli import script
